refactor(datasets): pipe /lines body parts through res.throttle, drop the bespoke write loop
res.endParts now pipes a Readable over a shifting generator through the SAME
res.throttle() bandwidth Transform the raw-download routes use: the pipe
provides the backpressure, res.throttle the token bucket + send-slot
queue-full teardown (slot released on the Throttle's 'close'), the streams'
high-water marks bound read-ahead to ~1MB, and the generator's shift()
releases each part as the pipe pulls it. This deletes the hand-rolled
backpressure layer from the previous commit (writeWithBackpressure,
drainOrAborted, writeParts, throttledSendParts + their unit tests) and
restores throttle-wait.ts and the wrapped res.end path byte-identical to
master — reverting the send-slot hold-time change that the drain-awaiting
loop had introduced on every throttled route (xlsx/shp/tiles/ODS keep
master's exact semantics; the new behavior is scoped to /lines).

Same observable behavior, verified by the same gates: /lines bandwidth
timing through the pipe, 304/HEAD parity, stream-source equivalence, ODS
compat, lint, typing ratchet at baseline.
