+++ tests/features/infra/rate-limiting.unit.spec.ts
@@ -1,8 +1,7 @@
 import { test } from '@playwright/test'
 import assert from 'node:assert/strict'
-import { EventEmitter } from 'node:events'
 import { TokenBucket } from '../../../api/src/misc/utils/token-bucket.ts'
-import { removeTokensOrAborted, tokenBucketFor, acquireSendSlot, releaseSendSlot, maxPendingSends, writeParts, throttledSendParts } from '../../../api/src/misc/utils/throttle-wait.ts'
+import { removeTokensOrAborted, tokenBucketFor, acquireSendSlot, releaseSendSlot, maxPendingSends } from '../../../api/src/misc/utils/throttle-wait.ts'
 
 // removeTokensOrAborted waits for the token bucket to grant `count` tokens, but bails out early if the
 // response/stream is torn down (its AbortSignal fires). Two production leaks are pinned here:
@@ -149,150 +148,5 @@ test.describe('send slots', () => {
       assert.equal(acquireSendSlot(tb), true, `slot ${i} should be granted`)
     }
     assert.equal(acquireSendSlot(tb), false)
-  })
-})
-
-// A hand-rolled PartsSink double: `full` makes write() report backpressure (return false), drain()
-// releases it. Deterministic — no reliance on real socket buffers.
-const fakeSink = () => {
-  const emitter = new EventEmitter()
-  return {
-    written: [] as Buffer[],
-    destroyed: false,
-    writableEnded: false,
-    full: false,
-    writableLength: undefined as number | undefined,
-    write (chunk: Buffer) { this.written.push(chunk); return !this.full },
-    once (ev: string, fn: () => void) { return emitter.once(ev, fn) },
-    off (ev: string, fn: () => void) { return emitter.off(ev, fn) },
-    drain () { emitter.emit('drain') },
-    get drainListeners () { return emitter.listenerCount('drain') }
-  }
-}
-
-const tick = () => new Promise(resolve => setImmediate(resolve))
-const parts3 = () => [Buffer.from('part-one'), Buffer.from('part-two'), Buffer.from('part-three')]
-
-// writeParts is the sequential-write half of the /lines partial-writes design: the response body is
-// sent as its accumulated parts (no final Buffer.concat) under backpressure, and the parts array is
-// consumed as it goes so already-sent parts become collectable DURING the send — the whole point.
-test.describe('writeParts', () => {
-  test('writes every part in order, consumes the array, resolves true', async () => {
-    const sink = fakeSink()
-    const parts = parts3()
-    const sig = countingSignal()
-    assert.equal(await writeParts(sink as any, parts, sig as any), true)
-    assert.equal(Buffer.concat(sink.written).toString(), 'part-onepart-twopart-three')
-    assert.equal(parts.length, 0, 'the array must be consumed (progressive release)')
-    assert.equal(sig.liveListeners, 0, 'no abort-listener residue')
-  })
-
-  test('honors backpressure: waits for drain between parts and releases them one by one', async () => {
-    const sink = fakeSink()
-    sink.full = true // every write reports a full socket buffer
-    const parts = parts3()
-    const sig = countingSignal()
-    const done = writeParts(sink as any, parts, sig as any)
-
-    await tick()
-    assert.equal(sink.written.length, 1, 'only one part may be in flight before the first drain')
-    assert.equal(parts.length, 3, 'nothing released yet — the write is not accepted')
-
-    sink.drain()
-    await tick()
-    assert.equal(sink.written.length, 2)
-    assert.equal(parts.length, 2, 'the accepted part must be released for collection')
-
-    sink.drain()
-    await tick()
-    sink.drain()
-    await tick()
-    assert.equal(await done, true)
-    assert.equal(Buffer.concat(sink.written).toString(), 'part-onepart-twopart-three')
-    assert.equal(parts.length, 0)
-    assert.equal(sink.drainListeners, 0, 'no drain-listener residue')
-  })
-
-  test('stops on abort while waiting for drain (client gone): resolves false, drops the tail', async () => {
-    const sink = fakeSink()
-    sink.full = true
-    const parts = parts3()
-    const sig = countingSignal()
-    const done = writeParts(sink as any, parts, sig as any)
-    await tick()
-    sig.abort()
-    assert.equal(await done, false, 'caller must not end() the response')
-    assert.ok(parts.length > 0, 'the tail is dropped, not written')
-    assert.equal(sink.written.length, 1)
-    assert.equal(sink.drainListeners, 0, 'the drain listener must be cleaned up on abort')
-  })
-
-  test('resolves false immediately on an already-destroyed sink', async () => {
-    const sink = fakeSink()
-    sink.destroyed = true
-    const sig = countingSignal()
-    assert.equal(await writeParts(sink as any, parts3(), sig as any), false)
-    assert.equal(sink.written.length, 0)
-  })
-
-  test('soft cap: a small buffered amount does not park on drain even when write() reports full', async () => {
-    // real responses expose writableLength; waiting for drain on EVERY refused 64KB write would
-    // lockstep the send to one event-loop round trip per part — below the ~1MB cap the loop keeps going
-    const sink = fakeSink()
-    sink.full = true
-    sink.writableLength = 1024 // far below the cap → never wait
-    const parts = parts3()
-    const sig = countingSignal()
-    assert.equal(await writeParts(sink as any, parts, sig as any), true, 'must complete without any drain event')
-    assert.equal(Buffer.concat(sink.written).toString(), 'part-onepart-twopart-three')
-    assert.equal(parts.length, 0)
-  })
-})
-
-// throttledSendParts drips the same parts under the bandwidth token bucket — plain res.write calls
-// would bypass the bucket entirely (a bandwidth-limit bypass), so the parity gate here is that a
-// multi-part body still takes at least the bucket-implied duration.
-test.describe('throttledSendParts', () => {
-  test('enforces the bandwidth limit across parts and preserves byte order', async () => {
-    // 20KB/s, burst 1 → bucket starts full with 20KB: a 30KB body must wait ~500ms for the last 10KB
-    const tb = tokenBucketFor(20000, 1)!
-    const parts = [Buffer.alloc(15000, 1), Buffer.alloc(10000, 2), Buffer.alloc(5000, 3)]
-    const expected = Buffer.concat(parts)
-    const written: Buffer[] = []
-    const sig = countingSignal()
-    const t0 = Date.now()
-    const sent = await throttledSendParts(tb, slice => { written.push(slice) }, parts, sig as any)
-    const elapsed = Date.now() - t0
-    assert.equal(sent, true)
-    assert.ok(Buffer.concat(written).equals(expected), 'bytes must be identical and in order')
-    assert.equal(parts.length, 0, 'parts are consumed as they are sent')
-    assert.ok(elapsed > 350, `a 30KB body over a 20KB burst + 20KB/s bucket should take ~500ms, took ${elapsed}ms`)
-  })
-
-  test('stops on abort mid-send: resolves false, keeps the unsent tail unconsumed', async () => {
-    const tb = tokenBucketFor(1000, 1)!
-    assert.ok(tb.bucket.tryTake(1000), 'drain the bucket so the send has to wait')
-    const parts = [Buffer.alloc(500, 1), Buffer.alloc(500, 2)]
-    const written: Buffer[] = []
-    const sig = countingSignal()
-    const done = throttledSendParts(tb, slice => { written.push(slice) }, parts, sig as any)
-    await sleep(20)
-    sig.abort()
-    assert.equal(await done, false)
-    assert.equal(written.length, 0, 'nothing was granted before the abort')
-    assert.equal(parts.length, 2)
-  })
-
-  test('awaits an async write (backpressure) before requesting the next slice tokens', async () => {
-    const tb = tokenBucketFor(100000, 4)!
-    const order: string[] = []
-    const sig = countingSignal()
-    const write = async (slice: Buffer) => {
-      order.push(`start-${slice.length}`)
-      await sleep(10)
-      order.push(`end-${slice.length}`)
-    }
-    assert.equal(await throttledSendParts(tb, write, [Buffer.alloc(10, 1), Buffer.alloc(20, 2)], sig as any), true)
-    assert.deepEqual(order, ['start-10', 'end-10', 'start-20', 'end-20'], 'writes must be strictly sequential')
   })
 })
